Reorganize RNN-T components in prototype module
