+++ examples/asr/librispeech_emformer_rnnt/pipeline_demo.py
@@ -3,7 +3,7 @@ from argparse import ArgumentParser
 
 import torch
 import torchaudio
-from torchaudio.prototype.rnnt_pipeline import EMFORMER_RNNT_BASE_LIBRISPEECH
+from torchaudio.prototype.rnnt import EMFORMER_RNNT_BASE_LIBRISPEECH
 
 
 def cli_main():
