+++ test/torchaudio_unittest/prototype/conformer_test_impl.py
@@ -1,5 +1,5 @@
 import torch
-from torchaudio.prototype import Conformer
+from torchaudio.prototype.nn import Conformer
 from torchaudio_unittest.common_utils import TestBaseMixin, torch_script
