+++ test/torchaudio_unittest/prototype/rnnt_decoder_test_impl.py
@@ -1,5 +1,5 @@
 import torch
-from torchaudio.prototype import RNNTBeamSearch, emformer_rnnt_model
+from torchaudio.prototype.rnnt import RNNTBeamSearch, emformer_rnnt_model
 from torchaudio_unittest.common_utils import TestBaseMixin, torch_script
