+++ torchaudio/prototype/rnnt/__init__.py
@@ -1,0 +1,14 @@
+from .rnnt import RNNT, emformer_rnnt_base, emformer_rnnt_model
+from .rnnt_decoder import Hypothesis, RNNTBeamSearch
+from .rnnt_pipeline import EMFORMER_RNNT_BASE_LIBRISPEECH, RNNTBundle
+
+
+__all__ = [
+    "Hypothesis",
+    "RNNT",
+    "RNNTBeamSearch",
+    "emformer_rnnt_base",
+    "emformer_rnnt_model",
+    "EMFORMER_RNNT_BASE_LIBRISPEECH",
+    "RNNTBundle",
+]
