+++ torchaudio/prototype/rnnt/rnnt_pipeline.py
@@ -8,7 +8,7 @@ from typing import Callable, List, Tuple
 import torch
 import torchaudio
 from torchaudio._internal import download_url_to_file, load_state_dict_from_url, module_utils
-from torchaudio.prototype import RNNT, RNNTBeamSearch, emformer_rnnt_base
+from torchaudio.prototype.rnnt import RNNT, RNNTBeamSearch, emformer_rnnt_base
 
 
 __all__ = []
@@ -157,7 +157,7 @@ class _SentencePieceTokenProcessor(_TokenProcessor):
 
 @dataclass
 class RNNTBundle:
-    """torchaudio.prototype.rnnt_pipeline.RNNTBundle()
+    """torchaudio.prototype.rnnt.RNNTBundle()
 
     Dataclass that bundles components for performing automatic speech recognition (ASR, speech-to-text)
     inference with an RNN-T model.
@@ -175,7 +175,7 @@ class RNNTBundle:
 
     Example
         >>> import torchaudio
-        >>> from torchaudio.prototype.rnnt_pipeline import EMFORMER_RNNT_BASE_LIBRISPEECH
+        >>> from torchaudio.prototype.rnnt import EMFORMER_RNNT_BASE_LIBRISPEECH
         >>> import torch
         >>>
         >>> # Non-streaming inference.
@@ -378,7 +378,7 @@ EMFORMER_RNNT_BASE_LIBRISPEECH = RNNTBundle(
 )
 EMFORMER_RNNT_BASE_LIBRISPEECH.__doc__ = """Pre-trained Emformer-RNNT-based ASR pipeline capable of performing both streaming and non-streaming inference.
 
-    The underlying model is constructed by :py:func:`torchaudio.prototypes.emformer_rnnt_base`
+    The underlying model is constructed by :py:func:`torchaudio.prototype.rnnt.emformer_rnnt_base`
     and utilizes weights trained on LibriSpeech using training script ``train.py``
     `here <https://github.com/pytorch/audio/tree/main/examples/asr/librispeech_emformer_rnnt>`__ with default arguments.
